Restructure component hierarchy
