+++ client/web/src/insights/components/insights-view-grid/components/backend-insight/components/drill-down-filters-action/DrillDownFiltersPanel.module.scss
@@ -17,15 +17,6 @@
     }
 }
 
-.filter-panel {
-    // class .dropdown-menu brings absolute position and this breaks
-    // the menu positioning in a portal because of it we reset position
-    // to static explicitly
-    position: static;
-    max-width: 32rem;
-    display: block;
-}
-
 .popover {
     display: block;
 }
