+++ client/web/src/insights/components/insights-view-grid/components/backend-insight/components/drill-down-filters-panel/DrillDownFiltersPanel.module.scss
@@ -1,30 +1,9 @@
-.drilldown-filters {
-    padding: 0.5rem 1rem;
+
+.filters-form {
+    max-width: 32rem;
+    display: block;
 }
 
-.header {
-    display: flex;
-}
-
-.regexp-field {
-    display: flex;
-
-    &__prefix {
-        flex-shrink: 0;
-        padding: 0.25rem 0.5rem;
-        height: 100%;
-        border: 1px solid var(--input-border-color);
-        background: var(--input-border-color);
-        border-top-left-radius: 4px;
-        border-bottom-left-radius: 4px;
-        color: var(--link-color);
-        display: flex;
-        align-items: center;
-        font-weight: normal;
-    }
-
-    &__input {
-        border-top-left-radius: 0;
-        border-bottom-left-radius: 0;
-    }
+.filters-view-creation {
+    max-width: 16rem;
 }
