+++ client/web/src/insights/components/insights-view-grid/components/backend-insight/components/drill-down-filters-panel/components/drill-down-filters-form/DrillDownFiltersForm.module.scss
@@ -1,0 +1,30 @@
+.drilldown-filters {
+    padding: 0.5rem 1rem;
+}
+
+.header {
+    display: flex;
+}
+
+.regexp-field {
+    display: flex;
+
+    &__prefix {
+        flex-shrink: 0;
+        padding: 0.25rem 0.5rem;
+        height: 100%;
+        border: 1px solid var(--input-border-color);
+        background: var(--input-border-color);
+        border-top-left-radius: 4px;
+        border-bottom-left-radius: 4px;
+        color: var(--link-color);
+        display: flex;
+        align-items: center;
+        font-weight: normal;
+    }
+
+    &__input {
+        border-top-left-radius: 0;
+        border-bottom-left-radius: 0;
+    }
+}
