+++ client/web/src/insights/components/insights-view-grid/components/backend-insight/components/drill-down-filters-panel/components/drill-down-filters-form/validators.ts
@@ -1,4 +1,4 @@
-import { Validator } from '../../../../../form/hooks/useField'
+import { Validator } from '../../../../../../../form/hooks/useField'
 
 export const validRegexp: Validator<string> = (value = '') => {
     if (value.trim() === '') {
