+++ client/web/src/insights/components/insights-view-grid/components/insight-card/InsightCard.module.scss
@@ -3,6 +3,7 @@
     flex-direction: column;
     padding: 0.5rem 1rem;
     cursor: default;
+    outline: none;
 
     &:focus,
     &:focus-within {
